🎨 fix import error with two classes in one file
