+++ lib/state.py
@@ -1,0 +1,9 @@
+class State():
+    FREE = 0
+    WALL = 1
+    HEALTHY = 2
+    INFECTED = 3
+    CURED = 4
+
+    ALLOWED_= [FREE, WALL, HEALTHY, INFECTED, CURED]
+
